% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_CI.R
\name{bootCounterfactuals}
\alias{bootCounterfactuals}
\title{bootCounterfactuals}
\usage{
bootCounterfactuals(result_t, t, mixture, weights, evgrid, grid)
}
\arguments{
\item{result_t}{A list containing the results of the DiSCo_CI_iter function}

\item{t}{The current time period}

\item{mixture}{Logical, indicating whether to use the mixture of distributions approach instead.
See Section 4.3. in \insertCite{gunsilius2023distributional;textual}{DiSCos}. This approach minimizes the distance between the CDFs
instead of the quantile functions, and is preferred for categorical variables. When working with such variables, one should
also provide a list of support points in the \code{grid.cat} parameter. When that is provided, this parameter is automatically set to TRUE. Default is FALSE.}

\item{grid}{Grid to recompute the CDF on if \code{mixture} option is chosen}
}
\value{
A list containing the bootstrapped counterfactuals
}
\description{
Function for computing the bootstrapped counterfactuals in the DiSCo method
}
\keyword{internal}
