% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checks}
\alias{checks}
\title{checks
Carry out checks on the inputs}
\usage{
checks(df, id_col.target, T0, M, G, num.cores, permutation)
}
\arguments{
\item{df}{Data frame or data table containing the distributional data for the target and control units. The data table should contain the following columns:
\itemize{
\item{\code{y_col} }{A numeric vector containing the outcome variable for each unit. Units can be individuals, states, etc., but they should be nested within a larger unit (e.g. individuals or counties within a state)}
\item{\code{id_col} }{A numeric vector containing the aggregate IDs of the units. This could be, for example, the state if the units are counties or individuals}
\item{\code{time_col} }{A vector containing the time period of the observation for each unit. This should be an increasing integer, and the first period should be equal to 1.}
}}

\item{id_col.target}{Variable indicating the name of the target unit, as specified in the id_col column of the data table.
This variable can be any type, as long as it is the same type as the id_col column of the data table.}

\item{M}{Integer indicating the number of control units to use in the DiSCo method. Default is 1000.}

\item{G}{Integer indicating the number of grid points for the grid on which the estimated functions are evaluated. Default is 1000.}

\item{num.cores}{Integer, number of cores to use for parallel computation. Default is 1 (sequential computation). If the \code{permutation} or \code{CI} arguments are set to TRUE, this can be very slow!}

\item{permutation}{logical, whether to use permutation or not}
}
\description{
checks
Carry out checks on the inputs
}
\examples{
df <- data.frame(id_col = c(1,1,1,2,2,2,3,3,3), t_col = c(1,2,3,1,2,3,1,2,3), y_col = c(1,1,0,1,1,0,1,1,0))
id_col.target <- 1
T0 <- 1
M <- 1000
G <- 1000
num.cores <- 1
permutation <- FALSE
checks(df, id_col.target, T0, M, G, num.cores, permutation)
}
\keyword{internal}
