% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_bc.R
\name{DiSCo_bc}
\alias{DiSCo_bc}
\title{Function for computing barycenters in the DiSCo method at every time period}
\usage{
DiSCo_bc(
  controls,
  controls.q,
  weights,
  evgrid = seq(from = 0, to = 1, length.out = 101)
)
}
\arguments{
\item{controls}{List with matrices of control distributions.}

\item{weights}{Vector of optimal synthetic control weights, computed using the DiSCo_weights_reg function.}

\item{evgrid}{Optional vector containing an evenly spaced grid on \link{0,1} on which the quantile function for the control units will be evaluated
By default, a grid of 100 points is used.}
}
\value{
The quantile function of the barycenter associated with the "weights" evaluated at the vector "evgrid"
}
\description{
Compute barycenters in the DiSCo method at every time period, as in Definition 1,
Step 4 in
This program takes in samples from the distributions of the controls and puts out
the associated barycenters and control group quantile functions.
The control distributions can be given in list form, where each list element contains a
vector of observations for the given control unit, in matrix form;
in matrix- each column corresponds to one unit and each row is one observation.
The list-form is useful, because the number of draws for each control group can be different.
}
\references{
\insertAllCited{}
}
