% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_CI.R
\name{DiSCo_CI}
\alias{DiSCo_CI}
\title{Confidence Intervals for the DiSCo quantile}
\usage{
DiSCo_CI(
  controls,
  bc,
  weights,
  mc.cores = 1,
  cl = 0.95,
  num.redraws = 500,
  evgrid = seq(from = 0, to = 1, length.out = 1001)
)
}
\arguments{
\item{cl}{a float specifying the confidence level.}

\item{num.redraws}{an integer specifying the number of redraws used in the bootstrap approach.}

\item{evgrid}{a vector of gridpoints used to evaluate quantile functions.}

\item{df_control}{a list or matrix:
in list form, each list element contains a vector of observations for the given control unit;
in matrix form, each column corresponds to one unit and each row is one observation.}

\item{df_target}{a vector of observations.}

\item{M}{an integer specifying the number of draws from the uniform distribution for approximating the integral.}

\item{solver}{a solver for the optimization problem; see \code{\link[CVXR]{installed_solvers}} in CVXR for more options.}

\item{w_t}{0 or a vector. By default, i.e. 0, \code{DSC_CI} calculates the confidence interval of a pre-treatment period,
weights of control units do not need to be specified. To calculate the confidence interval of a post-treatment period,
a vector of optimal weights (pre-calculated using pre-treatment observations) is needed.}

\item{graph}{\code{TRUE/FALSE}, indicating whether to output a plot of the confidence interval.}

\item{y_name}{a string for the title of the y-axis.}

\item{x_name}{a string for the title of the x-axis.}
}
\value{
\code{DSC_CI} returns a list containing the following components:
\itemize{
\item{\code{upper} }{a vector of the upper bound.}
\item{\code{lower} }{a vector of the lower bound.}
\item{\code{se} }{A vector of the standard errors of each counterfactual quantile estimate.}
\item{\code{bootmat} }{A matrix of the counterfactual quantile estimates for each bootstrap sample.}
}
}
\description{
Function for computing the confidence interval (at a specific period) in the DSC method
}
\examples{
#simulated data from Gaussian mixture
#ex_gmm() calls the simulated data
#details can be found by ??ex_gmm
DSC_CI(df_control=ex_gmm()$control, df_target=ex_gmm()$target, M=100, solver="SCS", w_t=0, cl=0.99, num.redraws=500, evgrid = seq(from=0, to=1, length.out=1001),graph=TRUE, y_name='y', x_name='x')
}
