% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_CI.R
\name{DiSCo_CI_iter}
\alias{DiSCo_CI_iter}
\title{DiSCo_CI_iter}
\usage{
DiSCo_CI_iter(
  redraw,
  controls,
  bc,
  weights,
  cl = 0.95,
  evgrid = seq(from = 0, to = 1, length.out = 1001),
  qmethod = NULL
)
}
\arguments{
\item{redraw}{Integer indicating the number of times the function has been called}

\item{controls}{List of control units}

\item{bc}{List of barycenters}

\item{weights}{Vector of optimal synthetic control weights, computed using the DiSCo_weights_reg function.}

\item{cl}{Confidence level for the confidence interval}

\item{evgrid}{Optional vector containing an evenly spaced grid on \link{0,1} on which the quantile function for the control units will be evaluated
By default, a grid of 1000 points is used.}

\item{qmethod}{Method for computing quantiles.}
}
\value{
The resampled counterfactual barycenter of the target unit
}
\description{
Function for computing the confidence intervals in the DiSCo method in a single period
}
\keyword{internal}
