% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_iter.R
\name{DiSCo_iter}
\alias{DiSCo_iter}
\title{DiSCo_iter}
\usage{
DiSCo_iter(
  yy,
  df,
  evgrid,
  id_col.target,
  M,
  G,
  T0,
  qmethod = NULL,
  q_min = 0,
  q_max = 1,
  ...
)
}
\arguments{
\item{yy}{Integer indicating the current year being processed.}

\item{df}{Data frame containing the data for both target and control units.}

\item{evgrid}{Vector containing an evenly spaced grid on \link{0,1} for evaluating quantile functions.}

\item{id_col.target}{String specifying the column name for the target unit's identifier.}

\item{M}{Integer specifying the number of iterations for optimization.}

\item{G}{Integer indicating the grid size for evaluation.}

\item{T0}{Integer indicating the first year of treatment as counted from 1.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{DiSCo_weights} }{Weights calculated using the DiSCo method.}
\item{\code{mixture} }{
\itemize{
\item{\code{weights} }{Optimal weights for the mixture approach.}
\item{\code{distance} }{Value of the objective function for the mixture approach.}
\item{\code{mean} }{Weighted mixture of the controls' CDFs.}
}
}
\item{\code{target} }{
\itemize{
\item{\code{cdf} }{Empirical CDF of the target.}
\item{\code{grid} }{Grid on which the quantile and CDF functions were evaluated.}
\item{\code{data} }{Original data for the target unit.}
\item{\code{quantiles} }{Quantiles for the target unit, evaluated on the specified grid.}
}
}
\item{\code{controls} }{
\itemize{
\item{\code{data} }{Original data for the control units.}
\item{\code{cdf} }{Empirical CDFs of the control units.}
\item{\code{id} }{IDs of the control units.}
\item{\code{quantiles} }{Quantiles for the control units, evaluated on the specified grid.}
}
}
\item{\code{controls.q} }{Quantiles for the control units, evaluated on the specified grid.}
}
}
\description{
This function implements the DiSCo method for a single time period, as well as the mixture of distributions approach.
}
\details{
This function is part of the DiSCo method, called for each time period. It calculates the optimal weights for the DiSCo method and the mixture of distributions approach for a single time period. The function processes data for both the target and control units, computes the quantile functions, and evaluates these on a specified grid. The function is designed to be used within the broader context of the DiSCo function, which aggregates results across multiple time periods.
}
