% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{DiSCo_per_rank}
\alias{DiSCo_per_rank}
\title{DiSCo_per_rank}
\usage{
DiSCo_per_rank(distt, distp, T0)
}
\arguments{
\item{distt}{List of squared Wasserstein distances between the target unit and the control units}

\item{distp}{List of squared Wasserstein distances between the control units}
}
\value{
List of p-values for each time period
}
\description{
This function ranks the squared Wasserstein distances and returns the p-values for each time period
}
\keyword{internal}
