% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_weights_reg.R
\name{DiSCo_weights_reg}
\alias{DiSCo_weights_reg}
\title{DiSCo_weights_reg}
\usage{
DiSCo_weights_reg(
  controls,
  target,
  M = 500,
  qmethod = NULL,
  simplex = FALSE,
  q_min = 0,
  q_max = 1
)
}
\arguments{
\item{controls}{List with matrices of control distributions}

\item{target}{Matrix containing the target distribution}

\item{M}{Optional integer, number of draws from the uniform distribution for approximating the integral. See section 3.1 in the paper.}

\item{qmethod}{Optional string, method for computing the quantile function. See section 3.1 in the paper.}

\item{simplex}{Optional boolean, if TRUE the weights are constrained to be non-negative and sum to one.}

\item{q_min}{Optional numeric, lower bound for the uniform distribution from which the quantile draws are taken.}

\item{q_max}{Optional numeric, upper bound for the uniform distribution from which the quantile draws are taken.}
}
\value{
Vector of optimal synthetic control weights
}
\description{
Function for obtaining the weights in the DiSCo method at every time period
}
\details{
Estimate the optimal weights for the distributional synthetic controls method.
solving the convex minimization problem in Eq. (2) in \insertCite{gunsilius2023distributional;textual}{DiSCo}..
using a regression of the simulated target quantile on the simulated control quantiles, as in Eq. (3),
\eqn{\underset{\vec{\lambda} \in \Delta^J}{\operatorname{argmin}}\left\|\mathbb{Y}_t \vec{\lambda}_t-\vec{Y}_{1 t}\right\|_2^2}.
For the constrained optimization we rely on the package pracma
the control distributions can be given in list form, where each list element contains a
vector of observations for the given control unit, in matrix form;
in matrix- each column corresponds to one unit and each row is one observation.
The list-form is useful, because the number of draws for each control group can be different.
The target must be given as a vector.
}
\references{
\insertAllCited{}
}
\keyword{internal}
