% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dube}
\alias{dube}
\title{Data from \insertCite{dube2019minimum}{DiSCos}}
\format{
\subsection{\code{dube}}{

A data frame with 652,870 rows and 3 columns:
\describe{
\item{id_col}{State FIPS}
\item{time_col}{Year}
\item{y_col}{\code{adj0contpov} variable in \insertCite{dube2019minimum;textual}{DiSCos}. Captures the distribution of equalized family income from wages and salary, defined as multiples of the federal poverty threshold.}
...
}
}
}
\usage{
dube
}
\description{
As used in the empirical application of \insertCite{gunsilius2023distributional;textual}{DiSCos}.
}
\keyword{datasets}
