% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGrid}
\alias{getGrid}
\title{getGrid}
\usage{
getGrid(target, controls, G)
}
\arguments{
\item{target}{A vector containing the data for the target unit}

\item{controls}{A list containing the data for the control units}

\item{G}{The number of grid points}
}
\value{
A list containing the following elements:
\itemize{
\item{grid.min}{The minimum value of the grid}
\item{grid.max}{The maximum value of the grid}
\item{grid.rand}{A vector containing the grid points}
\item{grid.ord}{A vector containing the grid points, ordered}
}
}
\description{
Set up a grid for the estimation of the quantile functions and CDFs
}
\examples{
set.seed(123)
target <- rnorm(100)
controls <- list(rnorm(100),rnorm(100),rnorm(100))
grid <- list(grid.min = NA, grid.max = NA, grid.rand = NA, grid.ord = NA)
grid[c("grid.min", "grid.max", "grid.rand", "grid.ord")] <- getGrid(target, controls, G)
plot(grid$grid.ord, type="l")
}
\keyword{internal}
