% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMixture.R
\name{getMixture}
\alias{getMixture}
\title{getMixture
Compute the Wasserstein barycenter of a set of distributions}
\usage{
getMixture(controls1, target, grid.min, grid.max, grid.rand)
}
\arguments{
\item{controls1}{List with matrices of control distributions.}

\item{target}{Vector of observations for the target unit.}

\item{grid.min}{Minimal value of the grid on which the CDFs are evaluated.}

\item{grid.max}{Maximal value of the grid on which the CDFs are evaluated.}

\item{grid.rand}{Random grid on which the CDFs are evaluated.}
}
\value{
A list containing the following elements:
\itemize{
\item{\code{cdf} }{A matrix containing the CDFs of the target and control units evaluated on the grid.}
\item{\code{distance.opt} }{The optimal value of the Wasserstein distance.}
\item{\code{mean} }{The optimal value of the Wasserstein barycenter.}
\item{\code{target.order} }{The target unit, ordered.}
\item{\code{weights.opt} }{The optimal weights.}
}
}
\description{
Compute the Wasserstein barycenter of a set of distributions
}
