% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.integer}
\alias{is.integer}
\title{Check if a vector is integer}
\usage{
is.integer(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if x is integer, FALSE otherwise
}
\description{
Check if a vector is integer
}
\keyword{internal}
