% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{mclapply.hack}
\usage{
mclapply.hack(..., verbose = FALSE, mc.cores = NULL)
}
\arguments{
\item{verbose}{Should users be warned this is hack-y? Defaults to FALSE.}
}
\description{
This function mimics forking (done with mclapply in Mac or Linux) for the
Windows environment.  Designed to be used just like mclapply.  Credit goes to
Nathan VanHoudnos.
}
\seealso{
mclapply
}
\keyword{internal}
