% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myquant}
\alias{myquant}
\title{Compute the empirical quantile function}
\usage{
myquant(X, q)
}
\arguments{
\item{X}{A vector containing the data}

\item{q}{A vector containing the quantiles}
}
\value{
A vector containing the empirical quantile function
}
\description{
Compute the empirical quantile function
}
\examples{
set.seed(123)
X <- rnorm(100)
q <- 0.1
myquant(X,q)
}
