% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseBoots}
\alias{parseBoots}
\title{parseBoots}
\usage{
parseBoots(CI_temp, cl)
}
\arguments{
\item{CI_temp}{A list containing the bootstrapped counterfactuals}

\item{cl}{The confidence level}
}
\value{
A list containing the confidence intervals for the quantiles and cdfs
}
\description{
Function for parsing the bootstrapped counterfactuals in the DiSCo method
}
\keyword{internal}
