% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCoTEA.R
\name{plotDistOverTime}
\alias{plotDistOverTime}
\title{Plot distribution of treatment effects over time}
\usage{
plotDistOverTime(
  cdf_centered,
  grid_cdf,
  t_start,
  t_max,
  n_per_window,
  CI,
  ci_lower,
  ci_upper,
  ylim = c(0, 1),
  xlim = NULL,
  cdf = TRUE,
  xlab = "Distribution Difference",
  ylab = "CDF",
  obsLine = NULL,
  savePlots = FALSE,
  plotName = NULL,
  lty = 1,
  lty_obs = 1,
  t_plot = NULL
)
}
\arguments{
\item{cdf_centered}{list of centered distributional statistics}

\item{grid_cdf}{grid}

\item{t_start}{start time}

\item{t_max}{maximum time}

\item{n_per_window}{number of windows per plot}

\item{CI}{logical indicating whether to plot confidence intervals}

\item{ci_lower}{lower confidence interval}

\item{ci_upper}{upper confidence interval}
}
\value{
plot of distribution of treatment effects over time
}
\description{
Plot distribution of treatment effects over time
}
