% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{print.permut}
\alias{print.permut}
\title{title print.permut}
\usage{
\method{print}{permut}(x, digits = 3)
}
\arguments{
\item{x}{Object of class permut}

\item{digits}{Number of digits to print}
}
\value{
Prints permutation test results
}
\description{
Print permutation test results
}
\examples{
print(x, digits=3)
}
\keyword{internal}
