% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCoTEA.R
\name{summary.DiSCoT}
\alias{summary.DiSCoT}
\title{summary.DiSCoT}
\usage{
\method{summary}{DiSCoT}(object, ...)
}
\arguments{
\item{object}{DiSCoT object}

\item{...}{Additional arguments}
}
\value{
summary of DiSCoT object
}
\description{
Summary of DiSCoT object
}
