% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{summary.permut}
\alias{summary.permut}
\title{title summary.permut}
\usage{
\method{summary}{permut}(x, digits = 3)
}
\arguments{
\item{digits}{Number of digits to print}

\item{object}{Object of class permut}
}
\value{
Prints permutation test results
}
\description{
Summarize permutation test results
}
\examples{
summary(x, digits=3)
}
