% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_bc.R
\name{DiSCo_bc}
\alias{DiSCo_bc}
\title{Function for computing barycenters in the DiSCo method at every time period}
\usage{
DiSCo_bc(controls.q, weights, evgrid = seq(from = 0, to = 1, length.out = 101))
}
\arguments{
\item{controls.q}{List with matrices of control quantile functions}

\item{weights}{Vector of optimal synthetic control weights, computed using the DiSCo_weights_reg function.}
}
\value{
The quantile function of the barycenter associated with the "weights" evaluated at the vector "evgrid"
}
\description{
Compute barycenters in the DiSCo method at every time period, as in Definition 1,
Step 4 in \insertCite{gunsilius2023distributional;textual}{DiSCos}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
