% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_CI.R
\name{DiSCo_CI}
\alias{DiSCo_CI}
\title{Confidence Intervals for the DiSCo quantile}
\usage{
DiSCo_CI(
  controls,
  bc,
  weights,
  mc.cores = 1,
  cl = 0.95,
  num.redraws = 500,
  evgrid = seq(from = 0, to = 1, length.out = 1001),
  qmethod = NULL
)
}
\arguments{
\item{controls}{A list containing the raw data for the control group}

\item{bc}{A list of barycenters for the control group}

\item{weights}{A vector of optimal weights}

\item{mc.cores}{Number of cores to use for parallelization}

\item{cl}{The confidence level for the confidence interval}

\item{num.redraws}{The number of bootstrap samples to draw}

\item{evgrid}{The grid of quantiles to evaluate the counterfactual quantile function}

\item{qmethod}{The method for computing the quantile function}
}
\value{
\code{DSC_CI} returns a list containing the following components:
\itemize{
\item{\code{upper} }{a vector of the upper bound.}
\item{\code{lower} }{a vector of the lower bound.}
\item{\code{se} }{A vector of the standard errors of each counterfactual quantile estimate.}
\item{\code{bootmat} }{A matrix of the counterfactual quantile estimates for each bootstrap sample.}
}
}
\description{
Function for computing the confidence intervals in the DiSCo method
}
