% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{DiSCo_per}
\alias{DiSCo_per}
\title{DiSCo_per}
\usage{
DiSCo_per(
  c_df,
  t_df,
  controls.q,
  target.q,
  T0,
  ww = 0,
  peridx = 0,
  evgrid = seq(from = 0, to = 1, length.out = 101),
  graph = TRUE,
  num_cores = 1,
  redo_weights = FALSE,
  weights = NULL
)
}
\arguments{
\item{c_df}{List with matrices of control distributions}

\item{t_df}{Matrix containing the target distribution}

\item{T0}{Integer indicating first year of treatment as counted from 1 (e.g, if treatment year 2002 was the 5th year in the sample, this parameter should be 5).}

\item{ww}{Optional vector of weights indicating the relative importance of each time period. If not specified, each time period is weighted equally.}

\item{peridx}{Optional integer indicating number of permutations. If not specified, by default equal to the number of units in the sample.}

\item{evgrid}{Optional vector containing an evenly spaced grid on \link{0,1} on which the quantile function for the control units will be evaulated.
By default, a grid of 100 points is used.}

\item{graph}{Boolean indicating whether to plot graphs}

\item{num_cores}{Integer, number of cores to use for parallel computation. Set to 1 by default (sequential computation), this can be very slow!}

\item{y_name}{Y axis label of the graph}

\item{x_name}{X axis label of the graph}
}
\value{
List of matrices containing synthetic time path of the outcome variable
for the target unit together with the time paths of the control units
}
\description{
Function to implement permutation test for Distributional Synthetic Controls
}
\details{
This program iterates through all units and computes the optimal weights on the other units
for replicating the unit of iteration's outcome variable, assuming that it is the treated unit.
See Algorithm 1 in \insertCite{gunsilius2023distributional;textual}{DiSCo} for more details.
}
\references{
\insertAllCited{}
}
