% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per_iter.R
\name{DiSCo_per_iter}
\alias{DiSCo_per_iter}
\title{DiSCo_per_iter}
\usage{
DiSCo_per_iter(
  c_df,
  c_df.q,
  t_df,
  T0,
  peridx,
  evgrid,
  idx,
  M = 1000,
  ww = 0,
  qmethod = NULL,
  per_q_min = 0,
  per_q_max = 1
)
}
\arguments{
\item{c_df}{List of control units}

\item{c_df.q}{List of quantiles of control units}

\item{t_df}{List of target unit}

\item{T0}{Index of last pre-treatment period}

\item{peridx}{Vector of indices of control units}

\item{evgrid}{Vector of evaluation grid points}

\item{idx}{Index of permuted target unit}

\item{ww}{Optional vector of weights. Default is uniform weights.}
}
\value{
List of squared Wasserstein distances between the target unit and the control units
}
\description{
This function performs one iteration of the permutation test
}
