% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCoTEA.R
\name{DiSCoTEA}
\alias{DiSCoTEA}
\title{DiSCoTEA}
\usage{
DiSCoTEA(disco, agg = "ATT", graph = TRUE)
}
\arguments{
\item{disco}{Output of the DiSCo function}

\item{agg}{String indicating the aggregation statistic to be used. Options are "ATT" (average treatment effect on the treated),}

\item{graph}{Boolean indicating whether to plot graphs. Default is TRUE.}
}
\description{
Distributional Synthetic Controls Treatment Effect Aggregator (DiSCoTEA)
}
\details{
This function takes in the output of the DiSCo_per function and computes aggregate treatment effect using a user-specified aggregation statistic. The default is the average treatment effect (ATE).
}
