% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ex_gmm}
\alias{ex_gmm}
\title{ex_gmm}
\usage{
ex_gmm(Ts = 2, num.con = 30, numdraws = 1000)
}
\arguments{
\item{Ts}{an integer indicating the number of time periods}

\item{num.con}{an integer indicating the number of control units}

\item{numdraws}{an integer indicating the number of draws}
}
\value{
\item{\code{target}}{a vector.}
\item{\code{control}}{a matrix.}
}
\description{
Example data for \code{DiSCo} command.
Returns simulated target and control that are mixtures of Gaussian distributions.
}
