% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGrid}
\alias{getGrid}
\title{getGrid}
\usage{
getGrid(target, controls, G)
}
\arguments{
\item{target}{A vector containing the data for the target unit}

\item{controls}{A list containing the data for the control units}

\item{G}{The number of grid points}
}
\value{
A list containing the following elements:
\itemize{
\item \code{grid.min} The minimum value of the grid
\item \code{grid.max} The maximum value of the grid
\item \code{grid.rand} A vector containing the grid points
\item \code{grid.ord} A vector containing the grid points, ordered
}
}
\description{
Set up a grid for the estimation of the quantile functions and CDFs
}
\keyword{internal}
