% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myQuant}
\alias{myQuant}
\title{Compute the empirical quantile function}
\usage{
myQuant(X, q, qmethod = NULL, ...)
}
\arguments{
\item{X}{A vector containing the data}

\item{q}{A vector containing the quantiles}
}
\value{
A vector containing the empirical quantile function
}
\description{
Compute the empirical quantile function
}
\keyword{internal}
