% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_CI.R
\name{parseBoots}
\alias{parseBoots}
\title{parseBoots}
\usage{
parseBoots(CI_temp, cl, q_disco, cdf_disco, q_obs, cdf_obs, uniform = TRUE)
}
\arguments{
\item{CI_temp}{A list containing the bootstrapped counterfactuals}

\item{cl}{The confidence level}

\item{q_disco}{The estimated quantiles around which to center}

\item{cdf_disco}{The estimated cdfs around which to center}

\item{q_obs}{The observed quantiles}

\item{cdf_obs}{The observed cdfs}

\item{uniform}{Whether to use uniform or pointwise confidence intervals}
}
\value{
A list containing the confidence intervals for the quantiles and cdfs
}
\description{
Function for parsing the bootstrapped counterfactuals in the DiSCo method
}
\keyword{internal}
