% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{permut}
\alias{permut}
\title{permut}
\usage{
permut(distp, distt, p_overall, J_1, q_min, q_max, plot)
}
\arguments{
\item{distp}{List of squared Wasserstein distances between the control units}

\item{distt}{List of squared Wasserstein distances between the target unit and the control units}

\item{p_overall}{Overall p-value}

\item{J_1}{Number of control units}

\item{q_min}{Minimum quantile}

\item{q_max}{Maximum quantile}

\item{plot}{ggplot object containing plot of squared Wasserstein distances over time for all permutations.}
}
\value{
A list of class permut, with the same elements as the input arguments.
}
\description{
Object to hold results of permutation test
}
