% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{print.permut}
\alias{print.permut}
\title{print.permut}
\usage{
\method{print}{permut}(x, ...)
}
\arguments{
\item{x}{Object of class permut}

\item{...}{Additional arguments}
}
\value{
Prints permutation test results
}
\description{
Print permutation test results
}
\keyword{internal}
