% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSCo_per.R
\name{summary.permut}
\alias{summary.permut}
\title{summary.permut}
\usage{
\method{summary}{permut}(object, ...)
}
\arguments{
\item{object}{Object of class permut}

\item{...}{Additional arguments}
}
\value{
Prints permutation test results
}
\description{
Summarize permutation test results
}
